move hub to Searches
